refactor: node version of mass-spring
