+++ mass-spring/README.md
@@ -27,6 +27,10 @@ It correctly bundles Solid in production mode and optimizes the build for the be
 The build is minified and the filenames include the hashes.<br>
 Your app is ready to be deployed!
 
+### `npm run test`
+
+Run the experiment in headless mode. Uses `tsx` to run the experiment with default parameters.
+
 ## Deployment
 
 Learn more about deploying your application with the [documentations](https://vitejs.dev/guide/static-deploy.html)
