+++ mass-spring/package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "test": "tsx src/script.ts"
   },
   "dependencies": {
     "rose": "^0.4.10",
@@ -14,6 +15,7 @@
     "vite-plugin-top-level-await": "^1.3.1"
   },
   "devDependencies": {
+    "tsx": "^4.7.2",
     "typescript": "^5.0.2",
     "vite": "^4.4.5",
     "vite-plugin-solid": "^2.7.0"
